pyproject.toml
[tool.poetry]
name = "net-shop"
version = "0.1.0"
description = ""
authors = ["alstacon <alstacon@gmail.com>"]
readme = "README.md"
packages = [{include = "net_shop"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
envparse = "^0.2.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.1"
django-extensions = "^3.2.1"


[tool.poetry.group.tests.dependencies]
pytest = "^7.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
